+++ src/shared/constants/index.ts
@@ -1,9 +1,1 @@
-export enum AppRoutes {
-  Home = "Home",
-  NotFound = "NotFound",
-}
-
-export const RoutePath: Record<AppRoutes, string> = {
-  [AppRoutes.Home]: "/",
-  [AppRoutes.NotFound]: "*",
-};
+export * from "./url";
