+++ src/shared/constants/url.ts
@@ -1,0 +1,9 @@
+export enum AppRoutes {
+  Home = "Home",
+  NotFound = "NotFound",
+}
+
+export const RoutePath: Record<AppRoutes, string> = {
+  [AppRoutes.Home]: "/",
+  [AppRoutes.NotFound]: "*",
+};
